Original change: https://android-review.googlesource.com/c/platform/system/keymint/+/2309530

+++ ta/src/tests.rs
@@ -11,7 +11,7 @@ use kmr_wire::{
 #[test]
 fn test_invalid_data() {
     // Cross-check that the hand-encoded invalid CBOR data matches an auto-encoded equivalent.
-    let rsp = error_rsp(ErrorCode::UnknownError);
+    let rsp = error_rsp(ErrorCode::UnknownError as i32);
     let rsp_data = rsp.into_vec().unwrap();
     assert_eq!(rsp_data, invalid_cbor_rsp_data());
 }
